==> 99_short_report.Rmd <==
---
title: "Simulation - Applied HRS Example"
author: "Peter Buto"
date: "`r Sys.Date()`"
output: 
  rmarkdown::pdf_document:
    fig_caption: yes
    includes:
      in_header: preamble-latex.tex
---

```{r setup, include=FALSE}

# Clear Environment
rm(list=ls())
gc()

# Libraries
library("tidyverse")
library("kableExtra")

# Directories
dir <- list(
  NA_table   = file.path("~/Documents/DP_HRS_Only/Tables/")
)

# IMPORT DATA/TABLES
input <- list(
  regresult = readRDS(file.path(dir$NA_table, "MainResults.rds")),
  figure1 = readRDS(file.path(dir$NA_table, "Figure1.RDS")),
  nTracker = readRDS(file.path(dir$NA_table, "../nTracker.RDS")),
  tbl1 = readRDS(file.path(dir$NA_table, "Table1_List.RDS")),
  tbl_dist = readRDS("../../DP_HRS_Only/Tables/DistanceTable.rds"),
  tbl_means = readRDS("../../DP_HRS_Only/Tables/Matched_Cohort_Means.rds"),
  tbl_sds = readRDS("../../DP_HRS_Only/Tables/Matched_Cohort_SDs.rds")
)


outcomes <- c("GENHEALTH", "SYSTOLIC_BP")

# Clean other Regression Outputs
for(out in outcomes){
  input$regresult[[out]] <- input$regresult[[out]] %>% 
    select(Instructions_NONE.R,
           Instructions_EXP.R,Instructions_OUT.R,
           Instructions_BOTH.R, Instructions_BOTH.R_TRUTH) %>%
    rename(`No Mediators` = Instructions_NONE.R,
           #`GS - No Mediators` = Instructions_NONE.R_TRUTH,
           `BMI` = Instructions_EXP.R,
           #`GS - BMI` = Instructions_EXP.R_TRUTH,
           !!out := Instructions_OUT.R,
           #!!paste0("GS - ",out) := Instructions_OUT.R_TRUTH,
           `Both Mediators` = Instructions_BOTH.R,
           `GS - Both Mediators` = Instructions_BOTH.R_TRUTH
    ) 

  # Sample Size Tracker
  input$nTracker[[out]] <- input$nTracker[[out]] %>% 
    rename(`No Mediators` = Instructions_NONE.R,
           `BMI` = Instructions_EXP.R,
           !!out := Instructions_OUT.R,
           `Both Mediators` = Instructions_BOTH.R) %>%
    select(`No Mediators`, `BMI`, out, `Both Mediators`)

# Table 1
  input$table1[[out]] <- cbind(input$tbl1[[out]]$Instructions_NONE.R, 
                               input$tbl1[[out]]$Instructions_EXP.R, 
                               input$tbl1[[out]]$Instructions_OUT.R, 
                               input$tbl1[[out]]$Instructions_BOTH.R)

# Table Distance
input$tbl_dist[[out]] <- input$tbl_dist[[out]] %>% 
  rename(`BMI` = Instructions_EXP.R,
         !!out := Instructions_OUT.R,
         `Both Mediators` = Instructions_BOTH.R)
}


# Diabetes Regression Output
input$regresult[["DIABETES"]] <- input$regresult[["DIABETES"]] %>% 
  select(Instructions_NONE.R, 
         Instructions_EXP.R, Instructions_OUT.R,Instructions_DIAB.R,
         #Instructions_BOTH.R, 
         Instructions_DIABA1c.R, Instructions_ALL3.R,
         Instructions_ALL3.R_TRUTH) %>%
  rename(`No Mediators` = Instructions_NONE.R,
         `BMI` = Instructions_EXP.R,
         `HbA1c` := Instructions_OUT.R,
         `Diabetes` = Instructions_DIAB.R,
         #`BMI & HbA1c` = Instructions_BOTH.R,
         `Diabetes and HbA1c` = Instructions_DIABA1c.R,
         `BMI, HbA1c, DM` = Instructions_ALL3.R,
         `Gold Standard` = Instructions_ALL3.R_TRUTH) 

# Diabetes Sample Size Tracker
input$nTracker[["DIABETES"]] <- input$nTracker[["DIABETES"]] %>% 
  rename(`None` = Instructions_NONE.R,
         `BMI` = Instructions_EXP.R,
         `HbA1c` := Instructions_OUT.R,
         `Diabetes` = Instructions_DIAB.R,
         #`BMI & HbA1c` = Instructions_BOTH.R,
         `Diabetes & HbA1c` = Instructions_DIABA1c.R,
         `BMI, HbA1c, DM` = Instructions_ALL3.R) %>%
  select(`None`, `BMI`, `HbA1c`, `Diabetes`, 
         #`BMI & HbA1c`, 
         `Diabetes & HbA1c`, `BMI, HbA1c, DM`)

# Diabetes Table 1
input$table1[["DIABETES"]] <- cbind(
  input$tbl1[["DIABETES"]]$Instructions_NONE.R,
  input$tbl1[["DIABETES"]]$Instructions_EXP.R,
  input$tbl1[["DIABETES"]]$Instructions_OUT.R,
  input$tbl1[["DIABETES"]]$Instructions_DIAB.R,
  #input$tbl1[["DIABETES"]]$Instructions_BOTH.R,
  input$tbl1[["DIABETES"]]$Instructions_DIABA1c.R,
  input$tbl1[["DIABETES"]]$Instructions_ALL3.R
)

# Diabetes Table Distance
input$tbl_dist[["DIABETES"]] <- input$tbl_dist[["DIABETES"]] %>% 
  rename(`BMI, HbA1c, DM` = Instructions_ALL3.R,
         `Diabetes` = Instructions_DIAB.R,
         `Diabetes & HbA1c` = Instructions_DIABA1c.R,
         `BMI` = Instructions_EXP.R,
         `HbA1c` := Instructions_OUT.R)
#         `HbA1c & BMI` = Instructions_BOTH.R)

# Order the images so that it makes somewhat more sense
input$figure1$DIABETES$data$Model <- c("BMI, HbA1c, DM", #"BMI & HbA1c", 
                                       "Diabetes", "Diabetes & HbA1c", "BMI", "None", "HbA1c")
input$figure1$DIABETES$data$Model <- factor(input$figure1$DIABETES$data$Model,
                          levels = c("None", "BMI", "HbA1c", "Diabetes", #"BMI & HbA1c", 
                                     "Diabetes & HbA1c", "BMI, HbA1c, DM"))

input$figure1$GENHEALTH$data$Model <- c("BOTH", "BMI", "NONE", "GENHEALTH")
input$figure1$GENHEALTH$data$Model <-
  factor(input$figure1$GENHEALTH$data$Model,
         levels = c("NONE", "BMI", "GENHEALTH", "BOTH"))

input$figure1$SYSTOLIC_BP$data$Model <- c("BOTH", "BMI", "NONE", "SYS_BP")
input$figure1$SYSTOLIC_BP$data$Model <-
  factor(input$figure1$SYSTOLIC_BP$data$Model,
         levels = c("NONE", "BMI", "SYS_BP", "BOTH"))

# ORder table means
input$tbl_means$DIABETES <- input$tbl_means$DIABETES %>% 
  select(Instructions_NONE.R_OLD,   Instructions_NONE.R_young,
         Instructions_EXP.R_OLD,    Instructions_EXP.R_young,
         Instructions_OUT.R_OLD,    Instructions_OUT.R_young,
         Instructions_DIAB.R_OLD,   Instructions_DIAB.R_young,
         #Instructions_BOTH.R_OLD,   Instructions_BOTH.R_young,
         Instructions_DIABA1c.R_OLD,Instructions_DIABA1c.R_young,
         Instructions_ALL3.R_OLD,   Instructions_ALL3.R_young)


```

\newpage 

# Objective

Apply the data pooling approach (from the main text) to a real dataset. 

# Methods

## Cohort

The Health and Retirement Study (HRS) is a longitudinal study that surveys a representative sample of ~20,000 people in the United States with over 20 years of follow-up. We selected participants whose first visit was in 1992 and had follow-up visits in 1994, 2008, and 2018. 

We randomly assigned participants into an "older" cohort and a "younger" cohort. The "older" cohort uses data only from 2008 and later, masking the participants' data prior to then; the "younger" cohort uses data from 2008 and earlier, masking their data collected afterwards. 

## Matching Process

The simulation follows the steps taken in the main text. Briefly, the "older" cohort's baseline characteristics (assumed to be static such as race/ethnicity or sex) are exact matched to the "younger" cohort's baseline characteristics. As some characteristics can vary over time, the "older" cohort's baseline characteristics are then matched to time-varying characteristics of the "younger" cohort (e.g. religion, job status, marital status). The pairs created from these steps were then ranked based on how similar they were based on a difference in a set of continuous variables (e.g. those with closer BMI values were considered more similar). 

## Outcomes

To demonstrate the effect of different matching strategies, we selected three different outcomes (`History of Diabetes`, `Self-reported general health (1: Excellent - 5: Poor)`, `Systolic Blood Pressure`). We selected 2018 as the outcome year to have both component cohorts approximate longitudinal cohorts, where the younger cohort would represent age 50s to 66 and the older cohort would represent ages 66 to 77.

## Matching Sets 

For each outcome, we matched on core demographic features: sex, race/ethnicity, and age (within 5 years of each other). We then selected a minimal set of matching variables (total 12 matching sets): a model without mediator variables, a model matching only on a mediating exposure, a model matching only on a mediating outcome, and a model matching on both a mediating exposure and outcome. The mediator values were taken from 2008, the overlap year between the cohorts.

We tracked the sample size at each step of the process in the following tables.

\newpage

### Diabetes Sample Size Tracker
```{r, echo=FALSE}
input$nTracker$DIABETES %>% 
  kable(caption = "Sample Size Tracker") %>%
  kable_styling("striped",
                latex_options = c("HOLD_position", "scale_down"))
```

### General Health Sample Size Tracker
```{r, echo=FALSE}
input$nTracker$GENHEALTH %>% 
  kable(caption = "Sample Size Tracker") %>%
  kable_styling("striped",
                latex_options = c("HOLD_position"))
```

### Systolic Blood Pressure Sample Size Tracker
```{r, echo=FALSE}
input$nTracker$SYSTOLIC_BP %>% 
  kable(caption = "Sample Size Tracker") %>%
  kable_styling("striped",
                latex_options = c("HOLD_position"))
```

\newpage

# Regression Results

We ran regressions in the pooled data with age, race/ethnicity, and sex as co-variates. For comparisons, we then ran the same analysis in those who were matched, labeled "Gold Standard". Results (Table 2) were largely similar between the two.

### Figure 1a. Diabetes Effect Estimates by Matching Set and Source
```{r, echo=FALSE}
suppressWarnings(print(input$figure1$DIABETES + 
                         theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))))
```

### Figure 1b. General Health Effect Estimates by Matching Set and Source
```{r, echo=FALSE}
suppressWarnings(print(input$figure1$GENHEALTH +
                         theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))))
```

### Figure 1c. Systolic BP Effect Estimates by Matching Set and Source
```{r, echo=FALSE}
suppressWarnings(print(input$figure1$SYSTOLIC_BP + 
                         theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))))
```

\newpage 

# QC

Mean values for each distance variables were largely similar in the younger and older cohorts across each of the four matching sets. The mean values were most similar when mediator variables were excluded. 

Each of the following tables present a different outcome (diabetes, general health, systolic blood pressure) grouped together by the matching set (no mediators, BMI, outcome, both mediators). 

Units are in the mediator values (percentages for HbA1c, Count [1-5] for Self-Rated Health, and mmHg for SBP).

```{r, echo=FALSE}
input$tbl_means$DIABETES %>% 
  kable(caption = "Diabetes - Comparison Mean Values: Old vs Young",
        "latex",
        booktabs = T,
        col.names = rep(c("Old","Young"),6)) %>%
  kable_styling("striped",
                latex_options = c("HOLD_position", "scale_down")) %>%
  add_header_above(c(" "=1, "No Mediators" = 2, "BMI" = 2, "HbA1c" = 2, "Diabetes" = 2, "Diabetes and HbA1c" = 2, "All Mediators" = 2))
```

```{r, echo=FALSE}
input$tbl_means$GENHEALTH %>% 
  kable(caption = "General Health - Comparison Mean Values: Old vs Young",
        "latex",
        booktabs = T,
        col.names = rep(c("Old","Young"),4)) %>%
  kable_styling("striped",
                latex_options = c("HOLD_position", "scale_down")) %>%
  add_header_above(c(" "=1, "Both Mediators" = 2, "GENHEALTH" = 2, "BMI" = 2, 
                     "No Mediators" = 2))
```

```{r, echo=FALSE}
input$tbl_means$SYSTOLIC_BP %>% 
  kable(caption = "Systolic BP - Comparison of Mean Values: Old vs Young",
        "latex",
        booktabs = T,
        col.names = rep(c("Old","Young"),4)) %>%
  kable_styling("striped",
                latex_options = c("HOLD_position", "scale_down")) %>%
  add_header_above(c(" "=1, "Both Mediators" = 2, "SYSTOLIC_BP" = 2, "BMI" = 2, 
                     "No Mediators" = 2))
```

\newpage

We can also present the average absolute difference within each matched pair (this is the same distance that is used to rank how similar the matches are). In nearly all cases, with exception of general health, the average difference increased when matching on both sets of variables. 

Units are in the mediator values (percentages for HbA1c, Count [1-5] for Self-Rated Health, and mmHg for SBP).

```{r, echo=FALSE}
input$tbl_dist$DIABETES %>%
  select(-Diabetes) %>%
  kable(caption = "Diabetes: Average difference in distance variables",
        "latex",
        booktabs = T) %>%
  kable_styling("striped",
                latex_options = c("HOLD_position"))
```

```{r, echo=FALSE}
input$tbl_dist$GENHEALTH %>%
  kable(caption = "General Health: Average difference in distance variables",
        "latex",
        booktabs = T) %>%
  kable_styling("striped",
                latex_options = c("HOLD_position"))
```

```{r, echo=FALSE}
input$tbl_dist$SYSTOLIC_BP %>%
  kable(caption = "Systolic Blood Pressure: Average difference in distance variables",
        "latex",
        booktabs = T) %>%
  kable_styling("striped",
                latex_options = c("HOLD_position"))
```


\newpage

In the following tables, we compared the demographics of those who _were_ matched to those that _were not_ matched among participants randomly assigned to the `Older` cohort.

```{r, echo=FALSE}
input$table1$DIABETES %>% 
  kable(caption = "Diabetes - Matched Vs Unmatched",
        "latex",
        booktabs = T,
        col.names = rep(c("Matched","Unmatched"),6)) %>%
  kable_styling("striped",
                latex_options = c("HOLD_position", "scale_down")) %>%
  add_header_above(c(" "=1, "No Mediators" = 2, "BMI" = 2, "HbA1c" = 2, 
                     "Diabetes" = 2, "Diabetes & HbA1c" = 2,
                     "Both Mediators" = 2))
```

```{r, echo=FALSE}
input$table1$GENHEALTH %>% 
  kable(caption = "General Health - Matched Vs Unmatched",
        "latex",
        booktabs = T,
        col.names = rep(c("Matched","Unmatched"),4)) %>%
  kable_styling("striped",
                latex_options = c("HOLD_position", "scale_down")) %>%
  add_header_above(c(" "=1, "No Mediators" = 2, "BMI" = 2, "GENHEALTH" = 2, 
                     "Both Mediators" = 2))
```

```{r, echo=FALSE}
input$table1$SYSTOLIC_BP %>% 
  kable(caption = "Systolic BP - Matched Vs Unmatched",
        "latex",
        booktabs = T,
        col.names = rep(c("Matched","Unmatched"),4)) %>%
  kable_styling("striped",
                latex_options = c("HOLD_position", "scale_down")) %>%
  add_header_above(c(" "=1, "No Mediators" = 2, "BMI" = 2, "SBP" = 2, 
                     "Both Mediators" = 2))
```

\newpage
\blandscape

## Appendix: Full Regression Results

```{r, echo=FALSE}
input$regresult$DIABETES %>% 
  kable(caption = "Regression Results",
        "latex",
        booktabs = T) %>%
  kable_styling("striped",
                latex_options = c("HOLD_position", "scale_down")) 
```

```{r, echo=FALSE}
input$regresult$GENHEALTH %>% 
  kable(caption = "Regression Results",
        "latex",
        booktabs = T) %>%
  kable_styling("striped",
                latex_options = c("HOLD_position", "scale_down"))
```

```{r, echo=FALSE}
input$regresult$SYSTOLIC_BP %>% 
  kable(caption = "Regression Results",
        "latex",
        booktabs = T) %>%
  kable_styling("striped",
                latex_options = c("HOLD_position", "scale_down"))
```

\elandscape


```{r export, include=FALSE}
input$nTracker$DIABETES  %>% write.csv("99_Outputs/Diabetes_N.csv")
input$nTracker$GENHEALTH %>% write.csv("99_Outputs/GenHealth_N.csv")
input$nTracker$SYSTOLIC_BP%>%write.csv("99_Outputs/SYSTOL_BP_N.csv")

input$regresult$DIABETES   %>% write.csv("99_Outputs/Diab_reg.csv")
input$regresult$GENHEALTH  %>% write.csv("99_Outputs/GenHealth_reg.csv")
input$regresult$SYSTOLIC_BP%>% write.csv("99_Outputs/SysBP_reg.csv")
```

```{r exportpng, include=FALSE}
input$figure1$DIABETES + 
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

ggsave(filename = "DIABETES.png", device="png", path="99_Outputs/")

input$figure1$GENHEALTH +
                         theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

ggsave(filename = "GENHEALTH.png", device="png", path="99_Outputs/")

input$figure1$SYSTOLIC_BP + 
                         theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

ggsave(filename = "SYS_BP.png", device="png", path="99_Outputs/")


```
